fix
rename


fix tests


fix proxy kernel multiple start


duplicated test
