+++ src/Microsoft.DotNet.Interactive/Connection/ConnectNamedPipe.cs
@@ -2,13 +2,11 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.CommandLine;
-using System.IO.Pipes;
-using System.Security.Principal;
 using System.Threading.Tasks;
 
 namespace Microsoft.DotNet.Interactive.Connection
 {
-    public class ConnectNamedPipe : ConnectKernelCommand<NamedPipeConnection>
+    public class ConnectNamedPipe : ConnectKernelCommand<NamedPipeKernelConnection>
     {
         public ConnectNamedPipe() : base("named-pipe",
                                          "Connects to a kernel using named pipes")
@@ -16,7 +14,7 @@ namespace Microsoft.DotNet.Interactive.Connection
             AddOption(new Option<string>("--pipe-name", "The name of the named pipe"));
         }
 
-        public override Task<Kernel> ConnectKernelAsync(NamedPipeConnection connection, KernelInvocationContext context)
+        public override Task<Kernel> ConnectKernelAsync(NamedPipeKernelConnection connection, KernelInvocationContext context)
         {
             return connection.ConnectKernelAsync();
         }
